+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ToPlannedTransformConverter.java
@@ -21,6 +21,8 @@ import org.gradle.execution.plan.Node;
 import org.gradle.execution.plan.ToPlannedNodeConverter;
 import org.gradle.initialization.DefaultPlannedTransform;
 import org.gradle.internal.taskgraph.CalculateTaskGraphBuildOperationType;
+import org.gradle.internal.taskgraph.NodeIdentity;
+import org.gradle.api.internal.artifacts.transform.TransformationIdentity;
 
 import java.util.List;
 import java.util.function.Function;
@@ -33,7 +35,7 @@ public class ToPlannedTransformConverter implements ToPlannedNodeConverter {
     }
 
     @Override
-    public CalculateTaskGraphBuildOperationType.TransformationIdentity getNodeIdentity(Node node) {
+    public TransformationIdentity getNodeIdentity(Node node) {
         TransformationNode transformationNode = (TransformationNode) node;
         return transformationNode.getNodeIdentity();
     }
@@ -44,7 +46,7 @@ public class ToPlannedTransformConverter implements ToPlannedNodeConverter {
     }
 
     @Override
-    public CalculateTaskGraphBuildOperationType.PlannedNode convert(Node node, Function<Node, List<? extends CalculateTaskGraphBuildOperationType.NodeIdentity>> findDependencies) {
+    public CalculateTaskGraphBuildOperationType.PlannedNode convert(Node node, Function<Node, List<? extends NodeIdentity>> findDependencies) {
         TransformationNode transformationNode = (TransformationNode) node;
         return new DefaultPlannedTransform(
             transformationNode.getNodeIdentity(),
