+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/ExecuteScheduledTransformationStepBuildOperationDetails.java
@@ -18,7 +18,6 @@ package org.gradle.api.internal.artifacts.transform;
 
 import com.google.common.collect.ImmutableMap;
 import org.gradle.internal.operations.trace.CustomOperationTraceSerialization;
-import org.gradle.internal.taskgraph.CalculateTaskGraphBuildOperationType;
 
 public class ExecuteScheduledTransformationStepBuildOperationDetails implements ExecuteScheduledTransformationStepBuildOperationType.Details, CustomOperationTraceSerialization {
 
@@ -37,7 +36,7 @@ public class ExecuteScheduledTransformationStepBuildOperationDetails implements 
     }
 
     @Override
-    public CalculateTaskGraphBuildOperationType.TransformationIdentity getTransformationIdentity() {
+    public TransformationIdentity getTransformationIdentity() {
         return transformationNode.getNodeIdentity();
     }
