+++ subprojects/enterprise-operations/src/main/java/org/gradle/api/internal/artifacts/transform/TransformationIdentity.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.internal.artifacts.transform;
+
+import org.gradle.api.internal.artifacts.component.ComponentIdentifier;
+import org.gradle.internal.taskgraph.NodeIdentity;
+
+import java.util.Map;
+
+public interface TransformationIdentity extends NodeIdentity {
+
+    /**
+     * Path of an included build.
+     */
+    String getBuildPath();
+
+    /**
+     * Project path within the build.
+     */
+    String getProjectPath();
+
+    ComponentIdentifier getComponentId();
+
+    Map<String, String> getSourceAttributes();
+
+    Class<?> getTransformType();
+
+    Map<String, String> getFromAttributes();
+
+    Map<String, String> getToAttributes();
+
+    long getTransformationNodeId();
+
+}
